Add EDITTIP to 'docRoot' & improved for 'vhRoot'.
[E][O]

Added EDITTIP to Item 'docroot' and improved EDITTIP for Item 'vhRoot'.
Both changes are to more directly point out that these setting
directories will not be created automatically by the server and must be
owned by the appropriate user.

scripts/text/common/Virtual_Hosts_Basic.hdoc
[Update] Item 'vhRoot': Updated EDITTIP on directory needing to
pre-exist to also mention that it must be user owned in EO.

scripts/text/common/Virtual_Hosts_General.hdoc
[Update] Item 'docRoot': Added EO namespaces.
Added EDITTIP to mention that this directory will not be automatically
created and that it needs to be user owned.